Refactor mockSdk exports and enhance Sidebar tests with defaultMockSdk integration. Added defaultMockSdk to provide consistent mock data for testing.
