+++ apps/iterable/test/mocks/index.ts
@@ -1,2 +1,2 @@
 export { mockCma } from './mockCma';
-export { mockSdk } from './mockSdk';
+export { mockSdk, defaultMockSdk } from './mockSdk';
